Accept FnMut closures, instead of just Fn
If the pathfinding APIs accept `FnMut` closures, this gives the user
more flexibility, as their closures can mutate some of their captured
state.  `FnMut` is a supertrait of `Fn`, so existing use cases are not
affected by this change.

The effect internally is that such closures can't be shared in multiple
places simultaneously, since calling an `FnMut` requires exclusive
access.  None of the current implementations have a problem with this;
they just needed to update to mutable bindings.  (That's an internal
detail, not an API change.)