Prioritize nodes using correct cost value
When prioritizing between equal estimated-cost nodes, seems it was incorrectly
using the parent node's cost so far. Instead, it's own cost so far should be
used.

Consider a simple graph:

    A ----> B
       \
        --> C

Where A is the start node, and B and C have equal estimated cost, but different
cost so far. The "cost" value would previously be 0 for both B and C in the
smallest cost holder, thus not ensuring that the lowest cost-so-far node is
picked for expansion first.

This commit also cleans up the now duplicated cost data, and the unnecessary
generic type.