Add a dummy dependency to fix a "cargo -Z minimal-version" problem
The `criterion` crate does not check that it can be configured with
`cargo -Z minimal-version` and thus indicates incorrect
(as in "not strict enough") dependencies in its `Cargo.toml`.

By forcing the use of a later version of `rayon`
(which is used by `criterion`), we prevent dragging in a
non-compiling version of `crossbeam-utils`.