yen: Implement Yen's algorithm
Yen's algorithm is a graph algorithm to find k-shortests
paths. It is useful to provide alternative paths besides the
shortest path.  An explanation of the algorithm can be found here
https://en.wikipedia.org/wiki/Yen's_algorithm.

This commit implements a basic version of yen's algorithm based on the
linked wikipedia article with a binary min-heap to speed up the candidate
selection. A basic set of tests has been implemented. No benchmarking
has been performed so far. The current time complexity is O(KN^3),
assumg the dijkstra implementation is O(N^2).