fix!(dfs): never visit the same node twice
All path combinations could be visited, thus leading to an explosion of
the number of possibilities. For DFS, this makes no sense as if we
encounter a state that has been previously seen, we know that there is
no possible way to reach a goal through this state or it would have
been encountered already.

However, this adds two new bounds on the node type: `Clone` and `Hash`.