fix(dijkstra)!: remove unneeded partial cost parameter
The `dijkstra_reach` function used a successors function which not only
took the node whose successors are seeked, but also the cost so far to
reach this node. This is an oversight, as this information can be
retrieved from the iterator returned by `dijkstra_reach`.

The user is free to stop iterating if a maximum cost is reached, so
there is no loss in functionality.